Refactor
Http Handler refactor:
- abstract class
- static method and attributes

Refactor main activity:
- rename main activity
- onClick method
