+++ .gitignore
@@ -83,3 +83,10 @@ lint/generated/
 lint/outputs/
 lint/tmp/
 # lint/reports/
+
+caches/
+.tmp/
+native/
+wrapper/
+daemon/
+android/
