+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
             android:name="com.google.android.actions"
             android:resource="@xml/network_security_config" />
         <activity
-            android:name=".MainActivity"
+            android:name=".R2"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
