+++ app/src/main/java/com/example/physio_plus_app/OkHttpHandler.java
@@ -1,30 +1,24 @@
 package com.example.physio_plus_app;
 
-import android.content.Context;
 import android.os.StrictMode;
-import android.util.Log;
-import android.widget.Toast;
 
 import java.io.IOException;
-import java.util.Iterator;
 
 import okhttp3.FormBody;
-import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.RequestBody;
-import okhttp3.Response;
 
-public class OkHttpHandler {
-    public OkHttpHandler() {
-        StrictMode.ThreadPolicy policy = new
-                StrictMode.ThreadPolicy.Builder().permitAll().build();
+public abstract class OkHttpHandler {
+    private static final StrictMode.ThreadPolicy policy = new StrictMode.ThreadPolicy.Builder().permitAll().build();
+    public static void insertData(Params params) throws IOException {
         StrictMode.setThreadPolicy(policy);
-    }
+        final String myIP = "192.168.1.6";
+        String url = "http://"+myIP+"/request.php";
 
-    public void insertData(String url, Params params) throws IOException {
         OkHttpClient client = new OkHttpClient().newBuilder().build();
         FormBody.Builder builder = new FormBody.Builder();
+
         if (params != null) {
             builder.add("title", params.getTitle());
             builder.add("desc", params.getDesc());
